Fix Bootstrap Variable import Error
Before importing the variables the functions have
to the imported, since they are used for some of
the variables. For now I have commented out the
imports, because we no longer will no longer be
using a custom font, but the native font stack
instead, to ensure optimal Font Rendering for
all devices.