Import Events Into Calender View  (Basic Setup) [#38894]
It is still debatable whether we use a unique event model for the calender events or if we use just one and let the existing models/event.ts implement CalenderEvent. One thing to consider: see comment in MetaEvent class. Imo wrapping the models/event elem in a specific calender event model is easier to maintain.