Update Whole Test Suite To Be Compatible With Sec
The functioning security (authentication) has some issues. I.e. obviously any request has to be authenticated from now on, which is not the case for the existing integration tests. Solution: Create a 'test' profile which does not use spring security and update test scenarios to use this profile (everything else in this profile the same as before except that the db is in-memory mode)