Fix Access Rights Of REST Calls
Consultation was changed to be a customer feature (previously it was exclusive for authenticated users which is wrong). Biggest change: differentiate between event posts. Each EventType has its own endpoint (method), so that it is possible to make specific access rights for each type. Previously we could not make a differnet policy for lets say rents and courses (first is customer level, yet the latter is only accesible by trainers and admins) -> now we can differntiate as needed.

Same changes may have to be applied for all the PUT operations for event.